fix: form request, show firstname, and formate number to currency

+++ database/migrations/2022_01_08_025029_create_employees_table.php
@@ -17,7 +17,7 @@ class CreateEmployeesTable extends Migration
         Schema::create('employees', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name_employee');
-            $table->smallInteger('total_salary');
+            $table->integer('total_salary');
             $table->timestamps();
         });
     }
